+++ tests/test_binnedloss.py
@@ -158,6 +158,12 @@ def test_binned_loss(weights, Loss, simultaneous):
     obs_binned = obs.with_binning(binning)
     test_values_binned = test_values.to_binned(obs_binned)
     binned_gauss = zfit.pdf.BinnedFromUnbinnedPDF(gaussian1, obs_binned, extended=scale)
+    binned_gauss_alt = gaussian2.to_binned(obs_binned, extended=scale)
+    counts = binned_gauss.counts()
+    counts_alt = binned_gauss_alt.counts()
+    assert np.allclose(counts, counts_alt)
+    binned_gauss_closure = binned_gauss.to_binned(obs_binned)
+    assert np.allclose(counts, binned_gauss_closure.counts())
     if simultaneous:
         obs_binned2 = obs.with_binning(14)
         test_values_binned2 = test_values.to_binned(obs_binned2)